+++ projects/rocthrust/test/test_reduce_into.cpp
@@ -1,6 +1,6 @@
 /*
- *  Copyright 2008-2013 NVIDIA Corporation
- *  Modifications Copyright© 2019-2026 Advanced Micro Devices, Inc. All rights reserved.
+ *  Copyright 2026 NVIDIA Corporation
+ *  Modifications Copyright© 2026 Advanced Micro Devices, Inc. All rights reserved.
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -40,10 +40,10 @@ using VectorTestsParams = ::testing::Types<
   Params<thrust::universal_vector<int>>,
   Params<thrust::universal_host_pinned_vector<int>>>;
 
-TESTS_DEFINE(ReduceTests, FullTestsParams);
-TESTS_DEFINE(ReduceIntegerTests, UnsignedIntegerTestsParams);
-TESTS_DEFINE(ReducePrimitiveTests, NumericalTestsParams);
-TESTS_DEFINE(ReduceVectorUnitTests, VectorTestsParams);
+TESTS_DEFINE(ReduceIntoTests, FullTestsParams);
+TESTS_DEFINE(ReduceIntoIntegerTests, UnsignedIntegerTestsParams);
+TESTS_DEFINE(ReduceIntoPrimitiveTests, NumericalTestsParams);
+TESTS_DEFINE(ReduceIntoVectorUnitTests, VectorTestsParams);
 
 template <typename T>
 struct plus_mod_10
@@ -54,59 +54,66 @@ struct plus_mod_10
   }
 };
 
-TYPED_TEST(ReduceVectorUnitTests, TestReduceSimple)
+TYPED_TEST(ReduceIntoVectorUnitTests, TestReduceIntoSimple)
 {
   using Vector = typename TestFixture::input_type;
   using T      = typename Vector::value_type;
 
   SCOPED_TRACE(testing::Message() << "with device_id= " << test::set_device_from_ctest());
 
-  Vector v{1, -2, 3};
+  Vector i{1, -2, 3};
+  Vector o(1);
 
   // no initializer
-  ASSERT_EQ(thrust::reduce(v.begin(), v.end()), 2);
+  thrust::reduce_into(i.begin(), i.end(), o.begin());
+  EXPECT_EQ(o[0], 2);
 
   // with initializer
-  ASSERT_EQ(thrust::reduce(v.begin(), v.end(), (T) 10), 12);
+  thrust::reduce_into(i.begin(), i.end(), o.begin(), T(10));
+  EXPECT_EQ(o[0], 12);
 }
 
-template <typename InputIterator>
-int reduce(my_system& system, InputIterator, InputIterator)
+template <typename InputIterator, typename OutputIterator>
+void reduce_into(my_system& system, InputIterator, InputIterator, OutputIterator output)
 {
   system.validate_dispatch();
-  return 13;
+  *output = 13;
 }
 
-TEST(ReduceTests, TestReduceDispatchExplicit)
+TEST(ReduceIntoTests, TestReduceIntoDispatchExplicit)
 {
   SCOPED_TRACE(testing::Message() << "with device_id= " << test::set_device_from_ctest());
 
-  thrust::device_vector<int> vec;
+  thrust::device_vector<int> i;
+  thrust::device_vector<int> o(1);
 
   my_system sys(0);
-  thrust::reduce(sys, vec.begin(), vec.end());
+  thrust::reduce_into(sys, i.begin(), i.end(), o.begin());
 
-  ASSERT_EQ(true, sys.is_valid());
+  EXPECT_TRUE(sys.is_valid());
+  EXPECT_EQ(o[0], 13);
 }
 
-template <typename InputIterator>
-int reduce(my_tag, InputIterator, InputIterator)
+template <typename InputIterator, typename OutputIterator>
+void reduce_into(my_tag, InputIterator, InputIterator, OutputIterator output)
 {
-  return 13;
+  *output = 13;
 }
 
-TEST(ReduceTests, TestReduceDispatchImplicit)
+TEST(ReduceIntoTests, TestReduceIntoDispatchImplicit)
 {
   SCOPED_TRACE(testing::Message() << "with device_id= " << test::set_device_from_ctest());
 
-  thrust::device_vector<int> vec;
+  thrust::device_vector<int> i;
+  thrust::device_vector<int> o(1);
 
-  int result = thrust::reduce(thrust::retag<my_tag>(vec.begin()), thrust::retag<my_tag>(vec.end()));
+  thrust::reduce_into(
+    thrust::retag<my_tag>(i.begin()), thrust::retag<my_tag>(i.end()), thrust::retag<my_tag>(o.begin()));
 
-  ASSERT_EQ(13, result);
+  EXPECT_EQ(o[0], 13);
 }
 
-TYPED_TEST(ReducePrimitiveTests, TestReduce)
+TYPED_TEST(ReduceIntoPrimitiveTests, TestReduceInto)
 {
   using T = typename TestFixture::input_type;
 
@@ -123,18 +130,20 @@ TYPED_TEST(ReducePrimitiveTests, TestReduce)
       thrust::host_vector<T> h_data =
         get_random_data<T>(size, get_default_limits<T>::min(), get_default_limits<T>::max(), seed);
       thrust::device_vector<T> d_data = h_data;
+      thrust::host_vector<T> h_result(1);
+      thrust::device_vector<T> d_result(1);
 
       T init = 13;
 
-      T h_result = thrust::reduce(h_data.begin(), h_data.end(), init);
-      T d_result = thrust::reduce(d_data.begin(), d_data.end(), init);
+      thrust::reduce_into(h_data.begin(), h_data.end(), h_result.begin(), init);
+      thrust::reduce_into(d_data.begin(), d_data.end(), d_result.begin(), init);
 
       test_equality(h_result, d_result, size - 1);
     }
   }
 }
 
-TYPED_TEST(ReduceTests, TestReduceMixedTypes)
+TYPED_TEST(ReduceIntoTests, TestReduceIntoMixedTypes)
 {
   using Vector = typename TestFixture::input_type;
   using T      = typename Vector::value_type;
@@ -144,20 +153,23 @@ TYPED_TEST(ReduceTests, TestReduceMixedTypes)
   // make sure we get types for default args and operators correct
   if constexpr (std::is_floating_point<T>::value)
   {
+    Vector int_input{1, 2, 3, 4};
+
     Vector float_input{1.5, 2.5, 3.5, 4.5};
 
     // float -> int should use using plus<int> operator by default
-    ASSERT_EQ(thrust::reduce(float_input.begin(), float_input.end(), (int) 0), 10);
-  }
-  else
-  {
-    Vector int_input{1, 2, 3, 4};
+    thrust::host_vector<int> int_output(1);
+    thrust::reduce_into(float_input.begin(), float_input.end(), int_output.begin(), int(0));
+    EXPECT_EQ(int_output[0], 10);
+
     // int -> float should use using plus<float> operator by default
-    ASSERT_EQ(thrust::reduce(int_input.begin(), int_input.end(), (float) 0.5), 10.5);
+    thrust::host_vector<float> float_output(1);
+    thrust::reduce_into(int_input.begin(), int_input.end(), float_output.begin(), float(0.5));
+    EXPECT_EQ(float_output[0], 10.5);
   }
 }
 
-TYPED_TEST(ReduceIntegerTests, TestReduceWithOperator)
+TYPED_TEST(ReduceIntoIntegerTests, TestReduceIntoWithOperator)
 {
   using T = typename TestFixture::input_type;
 
@@ -174,13 +186,15 @@ TYPED_TEST(ReduceIntegerTests, TestReduceWithOperator)
       thrust::host_vector<T> h_data =
         get_random_data<T>(size, get_default_limits<T>::min(), get_default_limits<T>::max(), seed);
       thrust::device_vector<T> d_data = h_data;
+      thrust::host_vector<T> h_result(1);
+      thrust::device_vector<T> d_result(1);
 
       T init = 3;
 
-      T cpu_result = thrust::reduce(h_data.begin(), h_data.end(), init, plus_mod_10<T>());
-      T gpu_result = thrust::reduce(d_data.begin(), d_data.end(), init, plus_mod_10<T>());
+      thrust::reduce_into(h_data.begin(), h_data.end(), h_result.begin(), init, plus_mod_10<T>());
+      thrust::reduce_into(d_data.begin(), d_data.end(), d_result.begin(), init, plus_mod_10<T>());
 
-      ASSERT_EQ(cpu_result, gpu_result);
+      ASSERT_EQ(h_result, d_result);
     }
   }
 }
@@ -200,7 +214,7 @@ struct plus_mod3
   }
 };
 
-TYPED_TEST(ReduceTests, TestReduceWithIndirection)
+TYPED_TEST(ReduceIntoTests, TestReduceIntoWithIndirection)
 {
   // add numbers modulo 3 with external lookup table
   using Vector = typename TestFixture::input_type;
@@ -212,12 +226,14 @@ TYPED_TEST(ReduceTests, TestReduceWithIndirection)
 
   Vector table{0, 1, 2, 0, 1, 2};
 
-  T result = thrust::reduce(data.begin(), data.end(), T(0), plus_mod3<T>(thrust::raw_pointer_cast(&table[0])));
+  Vector result(1);
 
-  ASSERT_EQ(result, T(1));
+  thrust::reduce_into(data.begin(), data.end(), result.begin(), T(0), plus_mod3<T>(thrust::raw_pointer_cast(&table[0])));
+
+  ASSERT_EQ(result[0], T(1));
 }
 
-TYPED_TEST(ReducePrimitiveTests, TestReduceCountingIterator)
+TYPED_TEST(ReduceIntoPrimitiveTests, TestReduceIntoCountingIterator)
 {
   using T = typename TestFixture::input_type;
 
@@ -229,87 +245,21 @@ TYPED_TEST(ReducePrimitiveTests, TestReduceCountingIterator)
 
   thrust::counting_iterator<T, thrust::host_system_tag> h_first   = thrust::make_counting_iterator<T>(0);
   thrust::counting_iterator<T, thrust::device_system_tag> d_first = thrust::make_counting_iterator<T>(0);
+  thrust::host_vector<T> h_result(1);
+  thrust::device_vector<T> d_result(1);
 
   T init = random_integer<T>();
 
-  T h_result = thrust::reduce(h_first, h_first + n, init);
-  T d_result = thrust::reduce(d_first, d_first + n, init);
+  thrust::reduce_into(h_first, h_first + n, h_result.begin(), init);
+  thrust::reduce_into(d_first, d_first + n, d_result.begin(), init);
 
   // we use ASSERT_NEAR because we're testing floating point types
   if (std::is_floating_point<T>::value)
   {
-    ASSERT_NEAR(h_result, d_result, h_result * 0.01);
+    ASSERT_NEAR(h_result[0], d_result[0], h_result[0] * 0.01);
   }
   else
   {
-    ASSERT_EQ(h_result, d_result);
-  }
-}
-
-void TestReduceWithBigIndexesHelper(int magnitude)
-{
-  thrust::constant_iterator<long long> begin(1);
-  thrust::constant_iterator<long long> end = begin + (1ll << magnitude);
-  ASSERT_EQ(_THRUST_STD::distance(begin, end), 1ll << magnitude);
-
-  long long result = thrust::reduce(thrust::device, begin, end);
-
-  ASSERT_EQ(result, 1ll << magnitude);
-}
-
-TEST(ReduceTests, TestReduceWithBigIndexes)
-{
-  SCOPED_TRACE(testing::Message() << "with device_id= " << test::set_device_from_ctest());
-
-  TestReduceWithBigIndexesHelper(30);
-#ifndef THRUST_FORCE_32_BIT_OFFSET_TYPE
-  TestReduceWithBigIndexesHelper(31);
-  TestReduceWithBigIndexesHelper(32);
-  TestReduceWithBigIndexesHelper(33);
-#endif
-}
-
-__global__ THRUST_HIP_LAUNCH_BOUNDS_DEFAULT void ReduceKernel(int const N, int* in_array, int* result)
-{
-  if (threadIdx.x == 0)
-  {
-    thrust::device_ptr<int> in_begin(in_array);
-    thrust::device_ptr<int> in_end(in_array + N);
-
-    result[0] = thrust::reduce(thrust::hip::par, in_begin, in_end);
-  }
-}
-
-TEST(ReduceTests, TestReduceDevice)
-{
-  SCOPED_TRACE(testing::Message() << "with device_id= " << test::set_device_from_ctest());
-
-  for (auto size : get_sizes())
-  {
-    SCOPED_TRACE(testing::Message() << "with size= " << size);
-
-    for (auto seed : get_seeds())
-    {
-      SCOPED_TRACE(testing::Message() << "with seed= " << seed);
-
-      thrust::host_vector<int> h_data   = get_random_data<int>(size, 0, 15, seed);
-      thrust::device_vector<int> d_data = h_data;
-      thrust::device_vector<int> d_output(1, 0);
-
-      int h_output = thrust::reduce(h_data.begin(), h_data.end());
-
-      hipLaunchKernelGGL(
-        ReduceKernel,
-        dim3(1, 1, 1),
-        dim3(128, 1, 1),
-        0,
-        0,
-        size,
-        thrust::raw_pointer_cast(&d_data[0]),
-        thrust::raw_pointer_cast(&d_output[0]));
-      HIP_CHECK(hipGetLastError());
-
-      ASSERT_EQ(h_output, d_output[0]);
-    }
+    ASSERT_EQ(h_result[0], d_result[0]);
   }
 }